+++ packages/account-tree-controller/CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add persistence support for user customizations ([#6221](https://github.com/MetaMask/core/pull/6221))
+  - New `accountGroupsMetadata` (of new type `AccountTreeGroupPersistedMetadata`) and `accountWalletsMetadata` (of new type `AccountTreeWalletPersistedMetadata`) state properties to persist custom names, pinning, and hiding states.
+  - Custom names and metadata survive controller initialization and tree rebuilds.
+  - Support for `lastUpdatedAt` timestamps for Account Syncing V2 compatibility.
+- Add setter methods for setting custom account group names, wallet names and their pinning state and visibility ([#6221](https://github.com/MetaMask/core/pull/6221))
 - Add `group.type` tag ([#6214](https://github.com/MetaMask/core/pull/6214))
   - This `type` can be used as a tag to strongly-type (tagged-union) the `AccountGroupObject`.
 - Add `group.metadata` metadata object ([#6214](https://github.com/MetaMask/core/pull/6214))
@@ -23,6 +28,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - This `type` can be used as a tag to strongly-type (tagged-union) the `AccountWalletObject`.
 - Defaults to the EVM account from a group when using `setSelectedAccountGroup` ([#6208](https://github.com/MetaMask/core/pull/6208))
   - In case no EVM accounts are found in a group (which should not be possible), it will defaults to the first account of that group.
+- Enhanced customization priority hierarchy in tree building ([#6221](https://github.com/MetaMask/core/pull/6221))
+  - Custom user names now take priority over default rule-generated names.
 
 ## [0.7.0]